requirements.txt
fastapi==0.104.1
uvicorn==0.24.0.post1
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
boto3==1.33.1
pydantic==2.5.2
python-dotenv==1.0.0
bcrypt==4.1.1
python-multipart==0.0.6
PyJWT==2.8.0
click==8.1.7
requests==2.31.0
rich==13.7.0
